requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.28.0
jinja2>=3.1.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0
reportlab>=4.0.0
openai>=1.0.0
python-multipart>=0.0.6
backports.zoneinfo>=0.2.1; python_version < "3.9"
qrcode[pil]>=7.4.2
pyngrok>=7.0.0


SQLAlchemy>=2.0
sse-starlette>=2.1.0
trafilatura>=1.6.0
readability-lxml>=0.8.1
pyyaml>=6.0
